// apps/web/src/lib/api/generated/types/getAuditLogsSortBy.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type GetAuditLogsSortBy = typeof GetAuditLogsSortBy[keyof typeof GetAuditLogsSortBy];


export const GetAuditLogsSortBy = {
  createdAt: 'createdAt',
  action: 'action',
  entityType: 'entityType',
} as const;